Move tests for catching errors into a separate test file.
